+++ CHANGELOG.md
@@ -242,6 +242,7 @@
   - 🟢 [`ts-no-unnecessary-template-expression`](https://github.com/gajus/eslint-plugin-jsdoc/blob/HEAD/docs/rules/ts-no-unnecessary-template-expression.md)
   - 🟢 [`ts-prefer-function-type`](https://github.com/gajus/eslint-plugin-jsdoc/blob/HEAD/docs/rules/ts-prefer-function-type.md)
   - 🔴 [`require-rejects`](https://github.com/gajus/eslint-plugin-jsdoc/blob/HEAD/docs/rules/require-rejects.md)
+  - Set [`allowIndentedSections: true`](https://github.com/gajus/eslint-plugin-jsdoc/blob/HEAD/docs/rules/check-indentation.md#allowindentedsections) for [`check-indentation`](https://github.com/gajus/eslint-plugin-jsdoc/blob/HEAD/docs/rules/check-indentation.md).
 - `@stylistic/eslint-plugin`: [4.2.0 → 5.6.1](https://github.com/eslint-stylistic/eslint-stylistic/compare/v4.2.0...v5.6.1)
 - `eslint-plugin-json-schema-validator`: [5.4.0 → 5.5.0](https://github.com/ota-meshi/eslint-plugin-json-schema-validator/compare/v5.4.0...v5.5.0)
 - `eslint-plugin-jsonc`: [2.20.0 → 2.21.0](https://github.com/ota-meshi/eslint-plugin-jsonc/compare/v2.20.0...v2.21.0)